notifications: stronger notification also when focused
Add a source-code config value which uses stronger notifications (sound,
tray, etc) while the window is focused - as if the window was not
focused, for notifications on buffers which are not the current one.

Hardcode-enable this stronger notification.

This should probably have a UI settings, but for now it does not.