notifications: workaround: notify also on first query message
The code was notifying only for buffers which are visible at the lists,
but when the buffer doesn't exist yet (first query from a user), the
notification code runs before the buffer is added to the view, which
ended up without a notification.

Work around this by excluding query buffers when checking for visibility
before notifying. We still have other checks for ignore lists etc, so it's
not a terrible workaround.