build: speller: enable also with KDE5
Previously the HAVE_SPELLER define which enables the speller config UI
was not set for KDE5, which probably was a bug. Now its set also for KDE5.

Not sure how/if KF5Sonnet can be used for spelling, but currently if
it's available but KDE is not (and the custom speller is not enabled)
then the speller config UI will not show up. Not sure if that's
different from the main Quassel branch.